// src/types/types.ts
export interface Flights {
  result: Result;
}

export interface Result {
  hello: string;
  flights:    FlightElement[];
  bestPrices: BestPrices;
}

export interface BestPrices {
  ONE_CONNECTION: Direct;
  DIRECT:         Direct;
}
// ...
export interface FlightElement {
  hasExtendedFare: boolean;
  flight:          FlightFlight;
  flightToken:     string;
}